backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.0
openpyxl==3.1.5
pandas>=2.2.3
openai==1.50.0
pydantic==2.9.0
python-dotenv==1.0.1
httpx>=0.27.0
aiofiles==24.1.0
gunicorn==23.0.0
